move the consul specific functionality to pkg/kp/rollstore/consul_store.go
